feat: model ops
